Move SmallRect to its own file and give it a constructor for a common use case.
Already named as the ultimate name of the type, which hasn't been renamed just yet.
